requirements.txt
ultralytics>=8.3.0
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
scikit-learn>=1.3
pandas>=2.1
numpy>=1.25
matplotlib>=3.8
seaborn>=0.13
Pillow>=10.0
tqdm>=4.66
streamlit>=1.33
opencv-python>=4.8
